requirements.txt
transformers>=4.41.2,<=4.46.1
datasets>=2.16.0,<=3.0.2
accelerate>=0.34.0,<=1.0.1
peft>=0.11.1,<=0.12.0
trl>=0.8.6,<=0.9.6
gradio>=4.0.0,<5.0.0
pandas>=2.0.0
scipy
einops
sentencepiece
tiktoken
protobuf
uvicorn
pydantic
fastapi
sse-starlette
matplotlib>=3.7.0
fire
packaging
pyyaml
modelscope
pypinyin
google-cloud-translate
peewee
faker
numpy<2.0.0
xformers
bitsandbytes
av
